requirements.txt
django
djangorestframework
drf_yasg
mysqlclient
Pillow
